requirements.txt
# FlowSlide Complete Dependencies
# Core web framework and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
aiohttp>=3.12.4

# Database libraries
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.19.0  # SQLite异步驱动
psycopg2-binary>=2.9.0  # PostgreSQL同步驱动
asyncpg>=0.29.0  # PostgreSQL异步驱动

# Document processing
markdown>=3.5.1
python-docx>=1.1.0
pypdf2>=3.0.1
requests>=2.31.0
chardet>=5.0.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
pillow>=10.0.0  # Image processing library (PIL)

# AI and LLM libraries
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
tiktoken>=0.5.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.3.27

# Search and research libraries
tavily-python>=0.3.0  # Web search API client

# Browser automation
playwright>=1.40.0

# Utilities
click>=8.0.0
python-dotenv>=1.0.0
rich>=13.0.0
